+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to this project will be documented in this file.
 
 ### Improvements to the user experience
 
+- The security badge now provides more detailed information. ([#2042](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/2042))
 - All users get muted upon talking stick completion ([#1954](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/merge_requests/1954))
 - Improve german translations for streaming platform([#1357](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/merge_requests/1357))
 - Improve A11Y of the coffee break timer announcement ([#1997](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1997))